added a fix for memory leak on 2gb ram
